Calculate range sizes
A hunk header contains two numbers indicating the size of the old and
new hunk range.  These ranges can be calculated directly from the list
of deltas included in the hunk.  By calculating the ranges ourself, we
both verify a hunk header when parsing it and generate correct headers
when producing a hunk.